src/main/test/java/com/aoc/y2023/ProblemDay11Test.java
package com.aoc.y2023;

import com.aoc.ProblemDayTest;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.fest.assertions.Assertions.assertThat;

public class ProblemDay11Test implements ProblemDayTest {

    private ProblemDay11 problem;

    @BeforeEach
    public void before() {
        problem = new ProblemDay11();
    }

    @Test
    @Override
    public void testSolveA() {
        assertThat(problem.solveA()).isEqualTo(9536038L);
    }

    @Test
    @Override
    public void testSolveB() {
        assertThat(problem.solveB()).isEqualTo(447744640566L);
    }

}
